requirements.txt
# --- Web 框架 ---
fastapi
uvicorn

# --- 環境變數 ---
python-dotenv

# --- HTTP 請求 (給 Spotify/Gemini 用) ---
httpx
requests

# --- AI 模型 ---
google-genai

# --- 資料庫驅動 (Postgres) ---
sqlalchemy
psycopg2-binary

# --- 資料庫驅動 (Redis) ---
redis

# --- 資料庫驅動 (MongoDB) ---
motor

# --- 登入驗證 (Auth) ---
python-jose[cryptography]
passlib
bcrypt==3.2.2
python-multipart

# --- 測試工具 ---
pytest
pytest-asyncio